# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD Laravel ",
        "description": "API para administrar multimples servicios y relaciones entre tablas",
        "contact": {
            "email": "jhonfredyaya04@gmail.com"
        },
        "license": {
            "name": "Repocitorio",
            "url": "https://github.com/JhonFac/Crud_Laravel.git"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/programs/page/{id}": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Obtiene todos los usuarios",
                "description": "Display a listing of the resource.",
                "operationId": "e21ebc41bed62fb6682cd3ef94378d0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Numero de la pagina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de programs"
                    }
                }
            }
        },
        "/api/programs": {
            "post": {
                "tags": [
                    "Programs"
                ],
                "summary": "Crea un nuevo usuario",
                "description": "Store a newly created resource in storage.",
                "operationId": "8d0c35c3c3311c4c49d8d937c9cc1c42",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente"
                    }
                }
            }
        },
        "/api/programs/{id}": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Obtiene el usuario por ID",
                "description": "Display the specified resource.",
                "operationId": "e8941a6be874de950a9f6b4a344c36cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Numero de la pagina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de usuarios"
                    }
                }
            },
            "put": {
                "tags": [
                    "Programs"
                ],
                "summary": "Actualiza un program existente",
                "description": "Update the specified resource in storage.",
                "operationId": "2dd064f5d7487723ee033289155ba71b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del program a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Programs"
                ],
                "summary": "Elimina un usuario existente",
                "description": "Remove the specified resource from storage.",
                "operationId": "fb27e0b0e540d33d30ae4b7986be6586",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuario eliminado exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
...
    },
...
}